add[myo-hsat]: Added PrivateManga Authorization and User Route Model Binding Authorize Middleware

+++ app/Http/Controllers/API/v1/PrivateMangaController.php
@@ -39,6 +39,8 @@ class PrivateMangaController extends Controller
         PrivateMangaFilterRequest $request,
         User $user)
     {
+
+        $this->authorize('create', PrivateManga::class);
         // Get the initial builder with search filtering
         $query = $filterRepository->filterBySearchKeyword($user,
             PrivateManga::class,
@@ -66,6 +68,7 @@ class PrivateMangaController extends Controller
      */
     public function store(StorePrivateMangaRequest $request)
     {
+        $this->authorize('create', PrivateManga::class);
         $privateManga = PrivateManga::create([
             'name'              => $request->name,
             'slug'              => Str::slug($request->name),
@@ -88,6 +91,7 @@ class PrivateMangaController extends Controller
      */
     public function show(User $user, PrivateManga $privateManga)
     {
+        $this->authorize('view', $privateManga);
 
         return $this->customResponse->success(PrivateMangaResource::make($privateManga));
     }
@@ -99,6 +103,7 @@ class PrivateMangaController extends Controller
         User $user,
         PrivateManga $privateManga)
     {
+        $this->authorize('update', $privateManga);
         $privateManga->update([
             'name'              => $request->name,
             'slug'              => Str::slug($request->name),
@@ -122,6 +127,7 @@ class PrivateMangaController extends Controller
      */
     public function destroy(User $user, PrivateManga $privateManga)
     {
+        $this->authorize('delete', $privateManga);
         $this->privateMangaGenreController->destroyMultiple($privateManga);
         $privateManga->delete();
